Add index.html and .gitignore files to project
Established the baseline of the project by adding a .gitignore file to clearly specify to Git which files and directories to exclude from version control. This is a crucial step to prevent unwanted files (like IDE settings, caches, etc.) from getting pushed into the repository. In addition, a basic index.html file was created to begin the structuring of the 'WebGPU Showcase' webpage, with a link to the first assignment already provided. This setup will facilitate the addition of more assignments later on.